refactor: extracted components from Report Builder preview and updated checkEmptyItems for safety
